feat: ajouté fonction get_machines_adjacentes

+++ src/lan/lan.h
@@ -14,7 +14,7 @@ ERREUR_CODE lan_deinit(Reseau** lan);
 ERREUR_CODE lan_nombre_machine(Reseau* lan, size_t* nbMachines);
 ERREUR_CODE lan_nombre_connexion(Reseau* lan, size_t *nbConnexions);
 
-size_t lan_get_machines_adjacentes(Reseau* lan, appareil** machines_adjacentes, appareil* machine);
+int lan_get_machines_adjacentes(Reseau* lan, appareil** machines_adjacentes, appareil* machine);
 ERREUR_CODE lan_ajout_machine(Reseau* lan, appareil* machine);
 ERREUR_CODE lan_ajout_connexion(Reseau* lan, Lien lien);
